Refactor Get many test case in database/models/user/__tests__/mongo.repository.test.ts
