+++ src/database/models/user/subclasses/confirmation-code.ts
